requirements.txt
Django==2.0.2
python-rrdtool